// ui/src/react-admin/generated/graphql-db-types-hetarchief.ts
export type GetTranslationsQuery = {
	__typename?: 'query_root';
	app_config: Array<{ __typename?: 'app_config'; name: string; value: any }>;
};
// ...
export const GetTranslationsDocument = `
    query getTranslations {
  app_config(where: {name: {_ilike: "translations_%"}}) {
    name
    value
  }
}
    `;